Added different new use cases around project and user entities

+++ src/MyFeedbackHub.Api/Features/Users/Delete/DeleteUserEndpoint.cs
@@ -1,7 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
+﻿using MyFeedbackHub.Api.Shared.Utils;
 using MyFeedbackHub.Api.Shared.Utils.Carter;
 using MyFeedbackHub.Application.Shared.Abstractions;
 using MyFeedbackHub.Application.Users.Delete;
+using MyFeedbackHub.Domain.Types;
 
 namespace MyFeedbackHub.Api.Features.Users.Delete;
 
@@ -9,27 +10,29 @@ public sealed class DeleteUserEndpoint : ICarterModule
 {
     public void AddRoutes(IEndpointRouteBuilder app)
     {
-        app.MapDelete("/users", async (
-            [FromQuery] Guid id,
+        app.MapDelete("/users/{id}", async (
+            Guid id,
             ICommandHandler<DeleteUserCommand> command,
+            IUserContext userContext,
             CancellationToken cancellationToken) =>
         {
+            if (userContext.Role == UserRoleType.Customer
+                || userContext.Role == UserRoleType.TeamMember)
+            {
+                return Results.Forbid();
+            }
+
             var result = await command.HandleAsync(new DeleteUserCommand(id), cancellationToken);
             if (result.HasFailed)
             {
-                return Results.Problem(
-                    statusCode: StatusCodes.Status400BadRequest,
-                    title: "User deletion failed",
-                    extensions: new Dictionary<string, object?>()
-                    {
-                        ["errorCode"] = result.ErrorCode
-                    });
+                return result.ToBadRequest("User deletion failure");
             }
 
             return Results.Ok();
         })
         .WithName("DeleteUser")
         .Produces(StatusCodes.Status200OK)
+        .Produces(StatusCodes.Status400BadRequest)
         .WithSummary("Delete user")
         .WithDescription("Delete user")
         .WithTags("User")
